requirements.txt
arabic-reshaper==2.1.1
asgiref==3.3.1
Brotli==1.0.9
click==7.1.2
dash==1.17.0
dash-core-components==1.13.0
dash-html-components==1.1.1
dash-renderer==1.8.3
dash-table==4.11.0
Django==3.1.3
django-bootstrap-form==3.4
django-plotly-dash==1.4.2
dpd-components==0.1.0
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Compress==1.8.0
future==0.18.2
html5lib==1.1
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.19.4
openpyxl==3.0.5
pandas==1.1.4
Pillow==8.0.1
pkg-resources==0.0.0
plotly==4.13.0
PyPDF2==1.26.0
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2020.4
reportlab==3.5.55
retrying==1.3.3
six==1.15.0
sqlparse==0.4.1
webencodings==0.5.1
Werkzeug==1.0.1
xhtml2pdf==0.2.5
xlwt==1.3.0

djangorestframework~=3.12.2